Don't check the bottom-right block for distance
Specifically, we don't want to treat the empty block as a block.
However, that "block" needs a representation in our matrix.
The representation we have chosen for it is to be a "0".
Hence, a 2x2 goal grid would be {{1,2},{3,0}} in our representation.

If, for example in the Hamming Distance metric, we were to check every
site for whether it holds the correct block, we would always arrive at
an even number for the distance, even though the correct Hamming
Distance for, say, the grid {{1,2},{0,3}} is 1, because only one "block"
is out of place.

Accordingly, only check n*n - 1 blocks, since if the empty space is not
where it should be, then at least 1 other block is not where it should
be either.

To keep with the above example, {{1,2},{0,3}} would return a Hamming
Distance of 1, since although the 1 and the 2 are in the correct
location, index 2 contains 0, although it should contain 3.  This new
implementation does not check index 3.