Always return true for isSolvable
At least until testing has been completed for known solvable boards.
After that, then a twin board can be implemented and iterated
in lockstep with the original board, to determine which of the
two are solveable, and thus whether the original board was the twin
which was solvable or not.