Fix isGoal comparison
Previously, the code was comparing two objects together to see,
essentially, if the objects were literally the same entity.
Unfortunately, this was not the desired behavior.

The desired behavior would be for the comparison to return
true if each entry within the block was the same as the
entry in the same index within the goal.

Accordingly, do an iterative comparison rather than
comparing objects.