Remove redundant isGoal checks
Since, for each board's initialization, we're already calculating
the Manhattan distance, and since that Manhattan distance will be
0 when and only when that board is the goal board, we can
simplify the isGoal() method merely to check whether this
board's Manhattan distance is 0.

If yes, then the board is the goal board.  If not, then the board
is not the goal board.