Initialize current as the smaller of the two boards.
Previously, current was set as the given board, then both the given
board and its twin were added to the MinPQ, then the minimum from the
PQ was deleted.

Unfortunately, it's not necessarily the case that the given board is the
minimum, and so the previous code allowed for the deletion of a board
that was different than current, on the first iteration.

Accordingly, ensure that current IS the minimum board by initializing
it as such, then entering the loop in which the minimum is deleted
and new neighbors are added until we've found the goal board.